:recycle: 코드 이동 및 호출 구조 변경
