+++ src/components/index.js
@@ -7,6 +7,5 @@ import ListOption from "./ListOption";
 import ListOptionGroup from "./ListOptionGroup";
 import {ExternalLink, ArrowLeft, ArrowDown, EditIcon} from "./Svg";
 import ClipboardCopy from "./ClipboardCopy";
-import CommonFunc from "./Common";
 
-export { Footer, Header, ListItem, ListOption, ListOptionGroup, ClipboardCopy, CommonFunc, ExternalLink, ArrowLeft, ArrowDown, EditIcon };
+export { Footer, Header, ListItem, ListOption, ListOptionGroup, ClipboardCopy, ExternalLink, ArrowLeft, ArrowDown, EditIcon };
